+++ appseed/draw2d_direct2d/draw2d_direct2d_graphics.cpp
@@ -1511,7 +1511,7 @@ namespace draw2d_direct2d
 
                keeper < ::draw2d::dib * > keep(&m_pdibAlphaBlend, NULL, m_pdibAlphaBlend, true);
 
-               return BitBlt(x, y, rectText.width(), rectText.height(), dib1->get_graphics(), 0, 0, SRCCOPY);
+               return BitBlt((int32_t) x, (int32_t) y, rectText.width(), rectText.height(), dib1->get_graphics(), 0, 0, SRCCOPY);
 
                /*BLENDFUNCTION bf;
                bf.BlendOp     = AC_SRC_OVER;
